correct migrations

+++ database/migrations/2023_10_31_113637_create_customers_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('adress');
             $table->string('email')->nullable();
             $table->string('phone_number')->unique();
-            $table->integer('ice')->nullable();
+            $table->bigInteger('ice')->nullable();
             $table->timestamps();
         });
     }
